SITE/requirements.txt
fastapi
uvicorn
pydantic
numpy
scikit-learn
pandas
sqlalchemy
passlib[bcrypt]
python-jose
python-socketio
joblib
